ossflow-dubbing/requirements.txt
# Dubbing-generator deps.
# TTS: ElevenLabs cloud (voice_id pre-registered in dashboard).
# XTTS-v2 eliminado — usaba coqui-tts + transformers + ~8-10GB de modelos.
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.6.0
torchaudio==2.6.0

# Audio stack — demucs (GPU) separa voz/fondo para el Flujo B v5.
demucs
pydub
soundfile>=0.12
# librosa powers the QA boundary report (spectral centroid + pyin F0).
# Soft dep of the qa module — if missing the sidecar falls back to MOS-only.
librosa>=0.10.1

# Cloud TTS backend.
elevenlabs>=1.0

# Local TTS backend (Piper, ONNX, CPU).
piper-tts

# Local TTS backend (Kokoro-82M, StyleTTS2, GPU/CPU).
kokoro>=0.9.4

# Direct dep for SynthesizerS2Pro HTTP client + S2ProServerManager health probe.
# Currently transitive via elevenlabs/fastapi but pinning makes the contract
# explicit so a future elevenlabs bump that drops httpx doesn't break s2pro.
httpx>=0.27

# Test-only
pytest>=7.4
pytest-asyncio>=0.23
